feat(corelib): Add corelib; start migrating grafana-incident-app to use it (#3761)
### BACKGROUND: Why is this change required?

closes https://github.com/grafana/incident/issues/3776

Now that we have multiple shipping components, we need to have a common
library for shared components and shared utility code.
This PR is the first step in that direction — adding a lib, targeting
this lib with backend code, and making
grafana-incident-app a dependency of it.

Longer term, we'll use it for common components and shared resources. It
should be grafana-core agnostic, meaning it should not
depend on being a grafana plugin. instead, it should work for any
project without a dependency on being a first-party grafana plugin.

### VALIDATION: How does reviewer validate changes?

no functionality in grafana-incident-app should be affected by this
change.
the only change is that we now have a lib and a dependency on it.

---------

Co-authored-by: Reem Tariq <reem.tariq@grafana.com>
Co-authored-by: twu <199651+twu@users.noreply.github.com>